moves files around
