+++ app/modules/crypto/index.ts
@@ -1,0 +1,2 @@
+export { decrypt } from './decrypt.server';
+export { encrypt } from './encrypt.server';
